move get_content to editing actions
